+++ tests/test.h
@@ -1,0 +1,18 @@
+#ifndef tests_h
+#define tests_h
+
+#include "CuTest.h"
+
+CuSuite* CompfilesGetSuite();
+
+CuSuite* fileUtilGetSuite();
+
+CuSuite *scannerUtilGetSuite();
+
+CuSuite* scannerGetSuite();
+
+CuSuite* tokensTestGetSuite();
+
+CuSuite* dfaTestGetSuite();
+
+#endif